+++ packages/vue/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.1](https://github.com/leanjs/leanjs/compare/@leanjs/vue@0.9.0...@leanjs/vue@0.9.1) (2022-12-07)
+
+**Note:** Version bump only for package @leanjs/vue
+
 # [0.9.0](https://github.com/leanjs/leanjs/compare/@leanjs/vue@0.7.2...@leanjs/vue@0.9.0) (2022-12-04)
 
 ### Bug Fixes