+++ packages/webpack/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.21.7](https://github.com/leanjs/leanjs/compare/@leanjs/webpack@0.21.6...@leanjs/webpack@0.21.7) (2022-12-07)
+
+**Note:** Version bump only for package @leanjs/webpack
+
 ## [0.21.6](https://github.com/leanjs/leanjs/compare/@leanjs/webpack@0.21.3...@leanjs/webpack@0.21.6) (2022-12-04)
 
 ### Bug Fixes